next

+++ src/main/kotlin/org/radarbase/kotlin/util/RadarUtilitiesImpl.kt
@@ -4,6 +4,8 @@ import org.apache.kafka.streams.KeyValue
 import org.apache.kafka.streams.KeyValue.pair
 import org.apache.kafka.streams.kstream.Window
 import org.apache.kafka.streams.kstream.Windowed
+import org.radarbase.kotlin.stream.phone.PhoneUsageCollector
+import org.radarbase.kotlin.stream.phone.TemporaryPackageKey
 import org.radarcns.kafka.AggregateKey
 import org.radarcns.kafka.ObservationKey
 import org.radarcns.stream.aggregator.AggregateList
@@ -11,8 +13,6 @@ import org.radarcns.stream.aggregator.NumericAggregate
 import org.radarcns.stream.aggregator.PhoneUsageAggregate
 import org.radarbase.stream.collector.AggregateListCollector
 import org.radarbase.stream.collector.NumericAggregateCollector
-import org.radarbase.java.stream.phone.PhoneUsageCollector
-import org.radarbase.java.stream.phone.TemporaryPackageKey
 
 class RadarUtilitiesImpl : RadarUtilities {
     override fun getWindowed(window: Windowed<ObservationKey>): AggregateKey {
@@ -40,7 +40,7 @@ class RadarUtilitiesImpl : RadarUtilities {
         getWindowedTuple(window),
         PhoneUsageAggregate(
             window.key().packageName,
-            collector.totalForegroundTime,
+            collector.totalForegroundTime.toDouble(),
             collector.timesTurnedOn,
             collector.categoryName,
             collector.categoryNameFetchTime
